requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pymongo==4.13.2
openai==1.98.0
python-dotenv==1.1.1
pyjwt==2.10.1
dnspython==2.7.0
gunicorn==22.0.0
